2000-07-21  Benjamin Kosnik  <bkoz@purist.soma.redhat.com>
	* src/stl-inst.cc: Change __sink_unused_warning to unsigned int
	for alpha.

	* testsuite/26_numerics/complex_inserters_extractors.cc (testall):
	Change return type to void.
	* testsuite/25_algorithms/lower_bound.cc: Add return values.
	* bits/locale_facets.h: Add return values for the generic cases.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@35183 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libstdc++-v3/src/stl-inst.cc
@@ -51,6 +51,6 @@ namespace std {
 
   template
     void
-    __sink_unused_warning<size_t>(size_t);
+    __sink_unused_warning<unsigned int>(unsigned int);
 
 } //std
